requirements.txt
flask
python-dotenv
pyjwt
pytest
coverage
waitress
